cleanup: Remove some dev deps and uv.lock
Many of the python dev dependencies are replaced by ruff. so rely on it
and the lsp server `ruff server` to catch and correct many things.

removed the hard version dependencies as this is intended as a template
repo, so the actual instance can lock these if they see a need.

removed `uv.lock` for the same reasons. when the user clones this repo
and runs `direnv allow` they will be prompted to create the lock file by
running `uv sync`

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>